r.resamp.stats: Fix quantile option being truncated to an integer (#7791)
The quantile variable is a double, but its value was parsed with atoi, so
any quantile below 1.0 became 0 and method=quantile returned the minimum
regardless of the requested quantile. Use atof, as r.neighbors, r.series
and r3.neighbors already do for the same option.

+++ raster/r.resamp.stats/main.c
@@ -304,7 +304,7 @@ int main(int argc, char *argv[])
         G_fatal_error(_("Unknown method <%s>"), parm.method->answer);
 
     if (menu[method].method == c_quant) {
-        quantile = atoi(parm.quantile->answer);
+        quantile = atof(parm.quantile->answer);
         closure = &quantile;
     }
